refactor: rename custom component folder
