+++ .github/workflows/release-versioning.yml
@@ -40,7 +40,7 @@ jobs:
         run: |
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
-          git add custom_components/dynamic_energy_calculator/manifest.json setup.cfg
+          git add custom_components/dynamic_energy_contract_calculator/manifest.json setup.cfg
           git commit -m "chore: bump to next version after release ${{ github.ref_name }}"
           git push origin HEAD:main
