+++ custom_components/dynamic_energy_contract_calculator/manifest.json
@@ -1,13 +1,13 @@
 {
-  "domain": "dynamic_energy_calculator",
+  "domain": "dynamic_energy_contract_calculator",
   "name": "Dynamic Energy Contract Calculator",
   "codeowners": ["@bvweerd"],
   "config_flow": true,
   "dependencies": [],
-  "documentation": "https://www.github.com/bvweerd/dynamic_energy_calculator",
+  "documentation": "https://www.github.com/bvweerd/dynamic_energy_contract_calculator",
   "homekit": {},
   "iot_class": "calculated",
-  "issue_tracker": "https://github.com/bvweerd/dynamic_energy_calculator/issues",
+  "issue_tracker": "https://github.com/bvweerd/dynamic_energy_contract_calculator/issues",
   "quality_scale": "silver",
   "requirements": [],
   "ssdp": [],
