+++ custom_components/dynamic_energy_contract_calculator/quality_scale.yaml
