+++ custom_components/dynamic_energy_contract_calculator/translations/en.json
