+++ tests/test_config_flow.py
@@ -1,10 +1,10 @@
 from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResultType
 
-from custom_components.dynamic_energy_calculator.config_flow import (
+from custom_components.dynamic_energy_contract_calculator.config_flow import (
     DynamicEnergyCalculatorConfigFlow,
 )
-from custom_components.dynamic_energy_calculator.const import (
+from custom_components.dynamic_energy_contract_calculator.const import (
     CONF_CONFIGS,
     CONF_SOURCE_TYPE,
     CONF_SOURCES,
